refactor initial
